Update CHANGELOG.md
Add text in the changelog to call out the updates on target runtime.  Someone could word it better; it's good to have it visible in the changelog in case users did not click on the migration guide right away.  

Currently the logging changes is not showing up in this changelog.. maybe it's already in the main branch?